+++ src/main/java/com/example/aplicacion/controllers/api_controllers/APILanguageController.java
@@ -1,4 +1,4 @@
-package com.example.aplicacion.controllers.apiControllers;
+package com.example.aplicacion.controllers.api_controllers;
 
 import com.example.aplicacion.entities.Language;
 import com.example.aplicacion.pojos.LanguageAPI;
