+++ src/main/resources/application.properties
@@ -10,6 +10,7 @@ spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
 spring.jpa.hibernate.ddl-auto=update
 spring.rabbitmq.host=rabbit
 spring.rabbitmq.port=5672
+logging.level.root=info
 docker.dockerfile.timeout=100
 problem.default.timeout=10
 problem.default.cores=1
