+++ src/test/java/com/example/aplicacion/integrationtest/api_controllers/TestAPIContestController.java
@@ -1,6 +1,6 @@
-package com.example.aplicacion.integrationtest.apiControllers;
+package com.example.aplicacion.integrationtest.api_controllers;
 
-import com.example.aplicacion.controllers.apiControllers.APIContestController;
+import com.example.aplicacion.controllers.api_controllers.APIContestController;
 import com.example.aplicacion.entities.Contest;
 import com.example.aplicacion.entities.Problem;
 import com.example.aplicacion.entities.Team;
